tests: Follow Git repository helper module
The tests separate Git command execution coverage from Git object IO
coverage, while repository helper cases remain in the command utility test
module.

The source helpers now live in utils.git_repository. Tests that import them
from utils.git fail at collection time and keep the test tree from
reflecting the module boundary.

This commit addresses that mismatch by moving repository helper tests into
tests/utils/test_git_repository.py and updating command tests to import
repository root lookup from git_stage_batch.utils.git_repository.

The final commit will add an architecture guard for the repository helper
boundary.
